Renamed algo files
