+++ include/pf/Algorithms.h
@@ -1,8 +1,8 @@
 #pragma once
-#include "algorithms/a_star.h"
-#include "algorithms/bds.h"
-#include "algorithms/bfs.h"
-#include "algorithms/dfs.h"
-#include "algorithms/dls.h"
-#include "algorithms/gbfs.h"
-#include "algorithms/iddfs.h"
+#include "algorithms/A_Star.h"
+#include "algorithms/BiDirectional.h"
+#include "algorithms/BestFirst.h"
+#include "algorithms/DepthFirst.h"
+#include "algorithms/DepthLimited.h"
+#include "algorithms/GreedyBestFirst.h"
+#include "algorithms/IterativeDeepeningDepthFirst.h"
