+++ include/pf/algorithms/BestFirst.h
