+++ include/pf/algorithms/DepthFirst.h
