+++ include/pf/algorithms/GreedyBestFirst.h
