+++ source/pf/algorithms/BiDirectional.cpp
@@ -1,5 +1,5 @@
-#include "pf/algorithms/bds.h"
-#include "pf/algorithms/algo_utils.h"
+#include "pf/algorithms/BiDirectional.h"
+#include "pf/algorithms/Utils.h"
 #include <algorithm>
 #include <queue>
 #include <unordered_set>
