+++ source/pf/algorithms/DepthFirst.cpp
@@ -1,5 +1,5 @@
-#include "pf/algorithms/dfs.h"
-#include "pf/algorithms/algo_utils.h"
+#include "pf/algorithms/DepthFirst.h"
+#include "pf/algorithms/Utils.h"
 #include <algorithm>
 #include <queue>
 #include <unordered_set>
